+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release (Unreleased)
 
-* `github.com/patrickcping/pingone-go-sdk-v2` : v0.14.4
+* `github.com/patrickcping/pingone-go-sdk-v2/management` : [v0.63.0](./management/CHANGELOG.md)
+  * **Breaking change** `GrantTypes` is now an optional field on application data models. [#495](https://github.com/patrickcping/pingone-go-sdk-v2/pull/495)
+  * **Feature** Add support for the Rate Limit Configuration API. [#493](https://github.com/patrickcping/pingone-go-sdk-v2/pull/493)
 * `github.com/patrickcping/pingone-go-sdk-v2/mfa` : [v0.24.0](./mfa/CHANGELOG.md)
   * **Enhancement** Added support for PingID in device authentication policies.